🔧 Fix Build Timeouts
Unlike the React version, a fresh version of Vue-Keyrune is built whenever it is required as a dependency, and as such can't be in it's own package.json. Fortunately the only reason why this was being done was to support CodeSandbox for the demo app, but it can have it's own package.json.